refactor tricubic library to follow common code conventions and add CMake support
