--- tricubic/ltricubic_utils.h
@@ -1,2 +1,0 @@
-int ijk2n(int i, int j, int k);
-void point2xyz(int p, int *x, int *y, int *z);
